requirements.txt
streamlit
discord.py
pandas